+++ src/app/_mbti/_result/components/header-nav-bar.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react"
 import { useState } from "react"
 
-import { TITLE_LIST } from "./constants"
+import { TITLE_LIST } from "../constants"
 import HeaderNavBar from "./header-nav-bar"
 
 const meta: Meta<typeof HeaderNavBar> = {
@@ -16,7 +16,7 @@ type Story = StoryObj<typeof HeaderNavBar>
 export const Primary: Story = {
   render: args => {
     // eslint-disable-next-line react-hooks/rules-of-hooks
-    const [selectedTitle, setSelectedTitle] = useState<(typeof TITLE_LIST)[number]>("팀조합")
+    const [selectedTitle, setSelectedTitle] = useState<(typeof TITLE_LIST)[number]>("팀결과")
 
     return (
       <div className="phone-size bg-gray-950 text-white">
